+++ src/app/models/account/payer-model.ts
@@ -1,0 +1,35 @@
+export class PayerModel{
+    PayerId: number;
+    PayerName: string;
+    BillToName:string;
+    Email: string;
+    Phone: string;
+    Fax: string;
+    NPI: string;
+    FedId: string;
+    ETIN: string;
+    Taxonomy: string;
+    MedicaidId: string;
+    IsActive: number;
+    CreatedOn: Date;
+    CreatedBy: number;
+    BillingId: number;
+    Contact: number;
+    Title: string;
+    Address: string;
+    city: string;
+    state: string;
+    zip: number;
+    url:string;
+    isActive:number;
+  NetTerms: string;
+  SenderId: string;
+  ClearingHouse: string;
+  LocationCode: string;
+  StateAggregator: string;
+  AggregatorPayerId: string;
+  LinkedPayer: string;
+
+
+
+}
